requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backports.tarfile==1.2.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==2.0.12
click==8.1.8
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
docutils==0.21.2
dotty-dict==1.3.1
ecdsa==0.19.1
Flask==2.0.1
Flask-Login==0.6.3
Flask-WTF==1.0.1
gitdb==4.0.12
GitPython==3.1.44
gotrue==1.3.1
google-auth==2.3.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
importlib_metadata==8.8.6
invoke==1.7.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.6.0
nh3==0.2.21
packaging==24.2
Pillow==8.3.2
pkginfo==1.12.1.2
postgrest==0.10.7
pyasn1==0.6.1
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-gitlab==3.15.0
python-jose==3.3.0
python-semantic-release==7.33.2
pywin32-ctypes==0.2.3
readme_renderer==44.0
realtime==1.0.6
requests==2.26.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
segno==1.5.2
semver==2.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.13.2
tqdm==4.67.1
twine==3.8.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==1.26.20
websockets==12.0
Werkzeug==2.0.1
WTForms==3.2.1
zipp==3.21.0
